# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Personal Finance API
  description: |
    API for the Personal Finance application. Provides endpoints for managing
    bank accounts, transactions, investments, market data, and notifications.

    ## Authentication
    Most endpoints require authentication via Supabase Auth. Include the session
    cookie or Authorization header with a valid JWT token.

    ## Response Format
    Successful responses follow this format:
    ```json
    {
      "success": true,
      "data": { ... },
      "meta": { "count": 10 }
    }
    ```

    Error responses follow this format:
    ```json
    {
      "success": false,
      "error": "Error message"
    }
    ```
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: /api
    description: API routes
...
paths:
...
  /teller/create-enrollment:
    post:
      tags: [Accounts]
      summary: Create bank enrollment
      description: Initiates a new Teller Connect enrollment
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Enrollment created
          content:
            application/json:
              schema:
                type: object
                properties:
                  enrollmentId:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /teller/exchange-token:
    post:
      tags: [Accounts]
      summary: Exchange Teller Connect token
      description: Exchanges a Teller Connect token for account access
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [accessToken, enrollmentId]
              properties:
                accessToken:
                  type: string
                enrollmentId:
                  type: string
                institutionName:
                  type: string
      responses:
        '200':
          description: Token exchanged successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
  /teller/sync:
    post:
      tags: [Accounts]
      summary: Sync account transactions
      description: Triggers a sync of transactions for a specific enrollment
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [enrollmentId]
              properties:
                enrollmentId:
                  type: string
      responses:
        '200':
          description: Sync completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  transactionCount:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase session cookie
    cronAuth:
      type: http
      scheme: bearer
      description: CRON_SECRET bearer token for scheduled jobs
    discordSignature:
      type: apiKey
      in: header
      name: X-Signature-Ed25519
      description: Discord interaction signature
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: false
              error:
                type: string
                example: "Unauthorized"
...